Create 1218. Longest Arithmetic Subsequence of Given Difference.java
mp = defaultdict(int): This line initializes a defaultdict with a default value of 0. A defaultdict is a dictionary-like object that provides a default value when accessing a non-existent key.

mx = 0: This line initializes the variable mx to 0. It will be used to keep track of the maximum length of the subsequence.

for i in arr:: This is a loop that iterates over each element i in the given arr array.

mp[i] = mp[i-d] + 1: This line updates the value in the mp dictionary for the key i. The value is calculated as the value at key i-d (if it exists) plus 1. This represents the length of the subsequence ending at i. By storing this length in the dictionary, we can build up the lengths of subsequences for different elements.

x = max(mp[i], mx): This line updates the maximum length mx by comparing the current length at key i in mp with the current maximum. If the length at key i is greater, it becomes the new maximum.

Finally, the function returns the maximum length of the subsequence.

Explanation from chatgpt. The solution is from me 😇.